Update AirVPN to use SERVER_HOSTNAMES
Gluetun internally maintains a giant list of all
AirVPN endpoints. Setting SERVER_COUNTRIES essentially
randomly picks a server that matches the country.
However, this list contains servers with IPv6 addresses
as well, which our Synology setup does not support.

Instead, we can use the hostname directly, which
round-robins to different servers every 5 minutes
based on load. There are different hostnames for
IPv4 and IPv6 so we bypass the problem entirely.